refactor(SPV-1056): add bsv.Satoshis type; move optional package higher.
