+++ engine/paymail/paymail_service_client_interface.go
@@ -4,13 +4,14 @@ import (
 	"context"
 
 	"github.com/bitcoin-sv/go-paymail"
+	"github.com/bitcoin-sv/spv-wallet/models/bsv"
 )
 
 // ServiceClient is a service that aims to make easier paymail operations.
 type ServiceClient interface {
 	GetSanitizedPaymail(addr string) (*paymail.SanitisedPaymail, error)
 	GetCapabilities(ctx context.Context, domain string) (*paymail.CapabilitiesPayload, error)
-	GetP2PDestinations(ctx context.Context, address *paymail.SanitisedPaymail, satoshis uint) (*paymail.PaymentDestinationPayload, error)
+	GetP2PDestinations(ctx context.Context, address *paymail.SanitisedPaymail, satoshis bsv.Satoshis) (*paymail.PaymentDestinationPayload, error)
 	GetP2P(ctx context.Context, domain string) (success bool, p2pDestinationURL, p2pSubmitTxURL string, format PayloadFormat)
 	StartP2PTransaction(alias, domain, p2pDestinationURL string, satoshis uint64) (*paymail.PaymentDestinationPayload, error)
 	GetPkiForPaymail(ctx context.Context, sPaymail *paymail.SanitisedPaymail) (*paymail.PKIResponse, error)
