+++ models/bsv/satoshis.go
@@ -1,0 +1,3 @@
+package bsv
+
+type Satoshis uint64
