+++ models/request/paymail/paymail_output.go
@@ -1,12 +1,13 @@
 package paymail
 
 import (
-	"github.com/bitcoin-sv/spv-wallet/models/request/optional"
+	"github.com/bitcoin-sv/spv-wallet/models/bsv"
+	"github.com/bitcoin-sv/spv-wallet/models/optional"
 )
 
 type Output struct {
 	To       string                 `json:"to"`
-	Satoshis uint                   `json:"satoshis"`
+	Satoshis bsv.Satoshis           `json:"satoshis"`
 	From     optional.Param[string] `json:"from,omitempty"`
 }
